Restructure modules and update tests
